+++ app/lib/page-encointer/new_bazaar/single_business/widgets/map_button.dart
@@ -4,10 +4,9 @@ import 'package:encointer_wallet/theme/custom/extension/theme_extension.dart';
 class MapButton extends StatelessWidget {
   const MapButton({
     super.key,
-    required this.name,
     required this.onPressed,
   });
-  final String name;
+
   final VoidCallback onPressed;
   @override
   Widget build(BuildContext context) {
@@ -20,7 +19,7 @@ class MapButton extends StatelessWidget {
           elevation: 3,
         ),
         child: Text(
-          name,
+          'Open Map',
           style: context.textTheme.displaySmall!.copyWith(fontSize: 16),
         ),
       ),
