+++ app/pubspec.yaml
@@ -130,8 +130,10 @@ flutter:
     - assets/images/assets/
     - assets/images/public/
     - assets/avatars/
+    - assets/icons/
     - assets/nctr_logo.svg
     - assets/nctr_logo_faces_only_thick.svg
+   
 
   # fonts:
   #   - family: Schyler
